+++ backends/arm/quantizer/TARGETS
@@ -10,7 +10,6 @@ runtime.python_library(
     ],
 )
 
-# Exposed through __init__.py
 runtime.python_library(
     name = "arm_quantizer",
     srcs = ["arm_quantizer.py"],
@@ -45,6 +44,7 @@ runtime.python_library(
     name = "arm_quantizer_utils",
     srcs = ["arm_quantizer_utils.py"],
     deps = [
+        "//caffe2:torch",
         ":quantization_config",
         "//pytorch/ao:torchao",
     ],
