+++ backends/cortex_m/test/misc/test_portable_int8.py
@@ -14,11 +14,9 @@ import executorch.exir
 import torch
 from executorch.backends.arm._passes import FoldAndAnnotateQParamsPass
 from executorch.backends.arm._passes.arm_pass_utils import get_first_fake_tensor
+from executorch.backends.arm.quantizer.arm_quantizer_utils import SharedQspecQuantizer
 from executorch.backends.arm.test.common import parametrize
-from executorch.backends.cortex_m.quantizer.quantizer import (
-    CortexMQuantizer,
-    SharedQspecQuantizer,
-)
+from executorch.backends.cortex_m.quantizer.quantizer import CortexMQuantizer
 from executorch.backends.cortex_m.test.tester import CortexMTester
 from executorch.backends.test.harness.stages import StageType
 from executorch.exir import EdgeCompileConfig
